+++ vignettes/createAutomaticPDFfunctions.R
@@ -1,3 +1,3 @@
- # pack <- "EHRtemporalVariability"
- # path <- find.package(pack)
- # system(paste(shQuote(file.path(R.home("bin"), "R")),"CMD", "Rd2pdf", shQuote(path)))
+ pack <- "EHRtemporalVariability"
+ path <- find.package(pack)
+ system(paste(shQuote(file.path(R.home("bin"), "R")),"CMD", "Rd2pdf", shQuote(path)))