+++ NEWS.md
@@ -1,3 +1,9 @@
+# EHRtemporalVariability 1.1.3
+
+* Fixed issue when a numerical input column has all its values as missing.
+* Fixed issue when showing warning at kernel density estimation with less than 4 values.
+* Added help to `IGTProjection` class
+
 # EHRtemporalVariability 1.1.2
 
 * Added `webshot` package to suggest list in DESCRIPTION to avoid Vignette compilation issues.